+++ gfx/drivers_context/x_ctx.c
@@ -814,7 +814,7 @@ static bool gfx_ctx_x_set_video_mode(void *data,
 
                         break;
                      }
-                     else if (versions[i][0] == g_major && versions[i][1] == g_minor)
+                     else if (versions[i][0] == (int)g_major && versions[i][1] == (int)g_minor)
                      {
                         /* The requested version was tried and is not supported, go ahead and fail since everything else will be lower than that. */
                         break;
